requirements.txt
absl-py==0.14.0
activate==0.2.1
altgraph==0.17
argon2-cffi==21.1.0
astroid==2.4.2
astunparse==1.6.3
attrs==21.2.0
auto-py-to-exe==2.8.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==4.1.0
bottle==0.12.19
bottle-websocket==0.2.9
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
clang==5.0
click==7.1.2
colorama==0.4.4
comtypes==1.1.9
cycler==0.10.0
datefinder==0.7.1
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
Eel==0.12.4
entrypoints==0.3
Flask==1.1.2
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gevent==21.1.2
gevent-websocket==0.10.1
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googletrans==3.0.0
greenlet==1.0.0
grpcio==1.40.0
gTTS==2.2.2
h11==0.9.0
h2==3.2.0
h5py==3.1.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.7.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonify==0.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.3
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lsHotword==1.0.2
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
MouseInfo==0.1.3
mysql-connector-python==8.0.25
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
pefile==2019.4.18
php==1.2.1
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///C:/Users/imswa/Downloads/downloads%20new/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
PyAutoGUI==0.9.52
pycparser==2.20
pydub==0.25.1
pygame==2.0.1
PyGetWindow==0.0.9
Pygments==2.10.0
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
pylint==2.6.0
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyRect==0.1.4
pyrsistent==0.18.0
PyScreeze==0.1.26
python-dateutil==2.8.1
pyttsx3==2.90
PyTweening==1.0.3
pytz==2021.1
pywhatkit==4.0
pywin32==300
pywin32-ctypes==0.2.0
pywinpty==1.1.4
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rfc3986==1.5.0
rsa==4.7.2
scikit-learn==1.0
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.8.0
six==1.15.0
sklearn==0.0
sniffio==1.2.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
traitlets==5.1.0
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
whichcraft==0.6.1
widgetsnbextension==3.5.1
wikipedia==1.4.0
wrapt==1.12.1
zope.event==4.5.0
zope.interface==5.4.0
